Remove vessel and bones files, remove usages localCellCount, totalCellCount, tlnc global variables. remove variables and functions connected with bones, vassel and stages

+++ src/utils.h
@@ -3,7 +3,7 @@
 #include<stdint.h>
 #include<string.h>
 void stopRun(int, const char*, const char*, int);
-void swap_Nbyte(char* data, int n, int m);
+void swap_Nbyte(char *data, uint64_t n, int m);
 void checkEndiannes();
 void getLocalRankAndSize(int rank, int size, int32_t * lrank,	int32_t *lsize);
 size_t getMemoryPerProcess(int32_t lsize);
